# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FallbackCondition": {
        "description": "A failure after which the gateway moves on to the route's next target.",
        "enum": [
          "status_5xx",
          "rate_limited",
          "timeout",
          "connection_error"
        ],
        "title": "FallbackCondition",
        "type": "string"
      },
      "FallbackPolicy": {
        "additionalProperties": false,
        "description": "The failures after which the gateway tries the next target. Empty means never.",
        "properties": {
          "on": {
            "items": {
              "$ref": "#/components/schemas/FallbackCondition"
            },
            "title": "On",
            "type": "array"
          }
        },
        "title": "FallbackPolicy",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetryPolicy": {
        "additionalProperties": false,
        "description": "How often a target is tried, the first try included, and how long to wait in between.\n\nThe wait before retry *n* is `min(backoff_ms * 2 ** (n - 1), max_backoff_ms)`, replaced by a\nprovider's `Retry-After` when `honour_retry_after` is set.",
        "properties": {
          "backoff_ms": {
            "default": 200,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Backoff Ms",
            "type": "integer"
          },
          "honour_retry_after": {
            "default": true,
            "title": "Honour Retry After",
            "type": "boolean"
          },
          "max_attempts": {
            "default": 2,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "max_backoff_ms": {
            "default": 2000,
            "maximum": 30000.0,
            "minimum": 0.0,
            "title": "Max Backoff Ms",
            "type": "integer"
          },
          "on_statuses": {
            "items": {
              "maximum": 599.0,
              "minimum": 400.0,
              "type": "integer"
            },
            "maxItems": 16,
            "title": "On Statuses",
            "type": "array"
          }
        },
        "title": "RetryPolicy",
        "type": "object"
      },
      "RouteConfig": {
        "additionalProperties": false,
        "description": "Which credentials serve the route's calls, in fallback order, and how failures are handled.\n\n`timeout_ms` is the whole budget for every attempt, wait and fallback of one call; for a\nstream it bounds the time to the first byte.",
        "properties": {
          "fallback": {
            "$ref": "#/components/schemas/FallbackPolicy"
          },
          "retry": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/Target"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Targets",
            "type": "array"
          },
          "timeout_ms": {
            "default": 60000,
            "maximum": 600000.0,
            "minimum": 1000.0,
            "title": "Timeout Ms",
            "type": "integer"
          }
        },
        "required": [
          "targets",
          "retry",
          "fallback"
        ],
        "title": "RouteConfig",
        "type": "object"
      },
      "RouteCreateIn": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RouteConfig"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ],
        "title": "RouteCreateIn",
        "type": "object"
      },
      "RouteOut": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RouteConfig"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "is_default",
          "config",
          "version",
          "updated_by",
          "updated_at"
        ],
        "title": "RouteOut",
        "type": "object"
      },
      "RouteRevertIn": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "RouteRevertIn",
        "type": "object"
      },
      "RouteUpdateIn": {
        "additionalProperties": false,
        "description": "A new config, and the version the client last read, which must still be the current one.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RouteConfig"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "config",
          "expected_version"
        ],
        "title": "RouteUpdateIn",
        "type": "object"
      },
      "RouteVersionOut": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RouteConfig"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "config",
          "updated_by",
          "created_at"
        ],
        "title": "RouteVersionOut",
        "type": "object"
      },
...
      "Target": {
        "additionalProperties": false,
        "description": "One provider credential the route can send a call to.",
        "properties": {
          "credential_id": {
            "format": "uuid",
            "title": "Credential Id",
            "type": "string"
          },
          "model_aliases": {
            "maxProperties": 32,
            "patternProperties": {
              "^[A-Za-z0-9._:/-]{1,128}$": {
                "pattern": "^[A-Za-z0-9._:/-]{1,128}$",
                "type": "string"
              }
            },
            "title": "Model Aliases",
            "type": "object"
          },
          "weight": {
            "default": 1,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "credential_id"
        ],
        "title": "Target",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "created_at",
          "email_verified"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Spanlight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/gateway/routes": {
      "get": {
        "description": "By name. Unpaginated: a project keeps a handful of routes.",
        "operationId": "list_routes_api_v1_projects__project_id__gateway_routes_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteOut"
                  },
                  "title": "Response List Routes Api V1 Projects  Project Id  Gateway Routes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List gateway routes",
        "tags": [
          "gateway"
        ]
      },
      "post": {
        "description": "Save the route as version 1. The project's first route becomes its default.\n\n`422` on `config.targets.<i>.credential_id` for a credential that is not the organization's;\n`409 ROUTE_NAME_TAKEN` for a name already used in the project.",
        "operationId": "create_route_api_v1_projects__project_id__gateway_routes_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a route",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/projects/{project_id}/gateway/routes/{route_id}": {
      "delete": {
        "description": "Delete the route and its history. `409 ROUTE_IN_USE` while a gateway key uses it.",
        "operationId": "delete_route_api_v1_projects__project_id__gateway_routes__route_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a route",
        "tags": [
          "gateway"
        ]
      },
      "get": {
        "operationId": "get_route_api_v1_projects__project_id__gateway_routes__route_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a route",
        "tags": [
          "gateway"
        ]
      },
      "put": {
        "description": "Replace the config, if the route is still at `expected_version`; bumps `version`.\n\n`409 ROUTE_VERSION_CONFLICT` when another save came first: reload and apply the change again.",
        "operationId": "update_route_api_v1_projects__project_id__gateway_routes__route_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save a new version of a route",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/projects/{project_id}/gateway/routes/{route_id}/default": {
      "post": {
        "description": "Gateway keys created without a route use the default one. Not a new version.",
        "operationId": "set_default_route_api_v1_projects__project_id__gateway_routes__route_id__default_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Make a route the default",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/projects/{project_id}/gateway/routes/{route_id}/revert": {
      "post": {
        "description": "Save the config of `version` again as a new version; the history is never rewritten.\n\n`404` for a version the route never had; `422` when that config names a credential deleted\nsince.",
        "operationId": "revert_route_api_v1_projects__project_id__gateway_routes__route_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRevertIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert a route",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/projects/{project_id}/gateway/routes/{route_id}/versions": {
      "get": {
        "description": "Every saved config, newest first. `404` for an unknown route.",
        "operationId": "list_route_versions_api_v1_projects__project_id__gateway_routes__route_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteVersionOut"
                  },
                  "title": "Response List Route Versions Api V1 Projects  Project Id  Gateway Routes  Route Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a route's versions",
        "tags": [
          "gateway"
        ]
      }
    },
...
  }
}